refactor: Decouple Dask and MPI parallel logic into a concurrent module
